Update to latest gradle and kotlin releases. Currently has an issue where the built modules still include the kotlin-stdlib jars.

+++ modules/urn.uk.chriscormack.netkernel.lighting.artnet/src/uk/chriscormack/netkernel/lighting/artnet/TickerState.kt
@@ -56,9 +56,9 @@ internal class TickerState(private val controller: ArtNetController, coroutineCo
         } else {
             hasFinished = false
             newValue = if (stepValue > 0) {
-                floor(startValue + (currentTickCount * stepValue)).toShort()
+                floor(startValue + (currentTickCount * stepValue)).toInt().toShort()
             } else {
-                ceil(startValue + (currentTickCount * stepValue)).toShort()
+                ceil(startValue + (currentTickCount * stepValue)).toInt().toShort()
             }
         }
